refactor(tunnel): lift remote (-R) forward engine into core
Twin of the local-forward lift (#2196): move RemoteForwarder + RemoteForwardConfig
into termihub_core::tunnel so the identical engine runs on the desktop or an
agent. Desktop re-exports both; desktop behaviour is unchanged. Adds an async
start_async constructor (the agent path is already async) and a bound_port
getter, plus a ReachableFrom::SshServer classification for the -R listen side,
which lives on the SSH server.

Part of #2185
