+++ core/src/tunnel/config.rs
@@ -26,6 +26,26 @@ pub struct LocalForwardConfig {
     pub remote_port: u16,
 }
 
+/// Configuration for remote port forwarding (ssh -R).
+///
+/// The listen socket binds on the **SSH server** (`remote_host:remote_port`);
+/// each incoming connection is forwarded to `local_host:local_port` resolved
+/// from the tunnel host's network. On an agent-hosted tunnel the "local" side is
+/// the agent — the target is reached from the agent, not the desktop. See
+/// `docs/concepts/future/stateless-ui-agent-tunnel-endpoints.html`.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct RemoteForwardConfig {
+    /// Address on the SSH server to bind.
+    pub remote_host: String,
+    /// Port on the SSH server to listen on.
+    pub remote_port: u16,
+    /// Local host to forward connections to (resolved from the tunnel host).
+    pub local_host: String,
+    /// Local port to forward connections to (resolved from the tunnel host).
+    pub local_port: u16,
+}
+
 /// Live traffic statistics for an active tunnel.
 #[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -56,6 +76,21 @@ mod tests {
         assert!(json.get("localHost").is_some());
         assert!(json.get("remotePort").is_some());
         let back: LocalForwardConfig = serde_json::from_value(json).unwrap();
+        assert_eq!(back, config);
+    }
+
+    #[test]
+    fn remote_forward_config_round_trips_camel_case() {
+        let config = RemoteForwardConfig {
+            remote_host: "0.0.0.0".to_string(),
+            remote_port: 8080,
+            local_host: "127.0.0.1".to_string(),
+            local_port: 3000,
+        };
+        let json = serde_json::to_value(&config).unwrap();
+        assert!(json.get("remoteHost").is_some());
+        assert!(json.get("localPort").is_some());
+        let back: RemoteForwardConfig = serde_json::from_value(json).unwrap();
         assert_eq!(back, config);
     }
