+++ src-tauri/src/tunnel/remote_forward.rs
@@ -1,246 +1,9 @@
-use std::sync::Arc;
+//! Remote (`ssh -R`) port-forward engine — re-exported from core (#2185).
+//!
+//! The engine was lifted into `termihub_core::tunnel::remote_forward` so the
+//! identical forwarder runs on the desktop or on a remote agent (S3, part of
+//! #2139). This module re-exports it so existing desktop call sites
+//! (`super::remote_forward::RemoteForwarder`) are unchanged. The unit tests for
+//! the relay path now live alongside it in core.
 
-use termihub_core::backends::ssh::handler::{
-    ForwardedChannelRegistry, IncomingChannel, SshSession,
-};
-
-use super::config::{RemoteForwardConfig, TunnelStats};
-use super::local_forward::ForwarderStats;
-
-/// Manages a remote port forwarding tunnel.
-///
-/// Registers a `tcpip_forward` on the SSH server and receives incoming channels
-/// via the [`ForwardedChannelRegistry`], then relays each to a local TCP target.
-pub struct RemoteForwarder {
-    task_handle: Option<tokio::task::JoinHandle<()>>,
-    stats: Arc<ForwarderStats>,
-    registry: ForwardedChannelRegistry,
-    bound_port: u32,
-    _session: SshSession,
-    /// Fires when the forward loop exits (the server-side channel source closed),
-    /// so the tunnel supervisor can observe forwarder death (#1243, GAP 2).
-    death: Option<tokio::sync::oneshot::Receiver<()>>,
-}
-
-impl RemoteForwarder {
-    /// Start a remote port forwarding tunnel.
-    ///
-    /// Calls `tcpip_forward` on the SSH server to request it to listen on
-    /// `config.remote_host:config.remote_port`. Incoming channels are routed
-    /// via the `registry` and forwarded to `config.local_host:config.local_port`.
-    pub fn start(
-        config: &RemoteForwardConfig,
-        session: SshSession,
-        registry: ForwardedChannelRegistry,
-    ) -> Result<Self, std::io::Error> {
-        let remote_host = config.remote_host.clone();
-        let remote_port = config.remote_port as u32;
-
-        let bound_port = tokio::task::block_in_place(|| {
-            tokio::runtime::Handle::current()
-                .block_on(session.tcpip_forward(remote_host.clone(), remote_port))
-        })
-        .map_err(|e| std::io::Error::other(e.to_string()))?;
-
-        tracing::info!(
-            "Remote forward listening on {}:{} (bound port {})",
-            config.remote_host,
-            config.remote_port,
-            bound_port
-        );
-
-        let (tx, rx) = tokio::sync::mpsc::unbounded_channel::<IncomingChannel>();
-        registry.lock().unwrap().insert(bound_port, tx);
-
-        let stats = Arc::new(ForwarderStats::new());
-        let stats_clone = Arc::clone(&stats);
-        let local_host = config.local_host.clone();
-        let local_port = config.local_port;
-
-        let (death_tx, death_rx) = tokio::sync::oneshot::channel();
-        let task_handle = tokio::spawn(async move {
-            let _death = death_tx;
-            Self::forward_loop(rx, &local_host, local_port, stats_clone).await;
-        });
-
-        Ok(Self {
-            task_handle: Some(task_handle),
-            stats,
-            registry,
-            bound_port,
-            _session: session,
-            death: Some(death_rx),
-        })
-    }
-
-    /// Get current tunnel statistics.
-    pub fn get_stats(&self) -> TunnelStats {
-        self.stats.to_tunnel_stats()
-    }
-
-    /// Take the forwarder-death receiver (once) for the tunnel supervisor.
-    pub fn take_death_signal(&mut self) -> Option<tokio::sync::oneshot::Receiver<()>> {
-        self.death.take()
-    }
-
-    /// Stop the forwarder and deregister from the channel registry.
-    pub fn stop(&mut self) {
-        if let Some(handle) = self.task_handle.take() {
-            handle.abort();
-        }
-        if let Ok(mut reg) = self.registry.lock() {
-            reg.remove(&self.bound_port);
-        }
-    }
-
-    async fn forward_loop(
-        mut rx: tokio::sync::mpsc::UnboundedReceiver<IncomingChannel>,
-        local_host: &str,
-        local_port: u16,
-        stats: Arc<ForwarderStats>,
-    ) {
-        while let Some(incoming) = rx.recv().await {
-            stats.increment_active();
-            let local_addr = format!("{}:{}", local_host, local_port);
-            let stats = Arc::clone(&stats);
-            tokio::spawn(async move {
-                relay_to_local(incoming.channel.into_stream(), &local_addr, &stats).await;
-                stats.decrement_active();
-            });
-        }
-    }
-}
-
-impl Drop for RemoteForwarder {
-    fn drop(&mut self) {
-        self.stop();
-    }
-}
-
-/// Relay a single incoming forwarded channel to the configured local target.
-///
-/// Generic over the channel byte stream so the relay/stats-accounting logic is
-/// unit-testable with an in-memory stream — the production caller passes the
-/// russh `ChannelStream` from `incoming.channel.into_stream()` (#2044).
-async fn relay_to_local<S>(mut channel_stream: S, local_addr: &str, stats: &ForwarderStats)
-where
-    S: tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin,
-{
-    let mut tcp = match tokio::net::TcpStream::connect(local_addr).await {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!(
-                "Remote forward: failed to connect to local {}: {}",
-                local_addr,
-                e
-            );
-            return;
-        }
-    };
-
-    if let Ok((sent, received)) = tokio::io::copy_bidirectional(&mut tcp, &mut channel_stream).await
-    {
-        stats.add_bytes_sent(sent);
-        stats.add_bytes_received(received);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    //! Unit tests for the remote forwarder's per-channel relay to a local
-    //! target, driven with an in-memory duplex standing in for the SSH channel
-    //! plus a loopback echo server as the local target (#2044). The
-    //! `tcpip_forward` request itself needs a live SSH session and is covered by
-    //! the integration lanes; here we lock in the relay + stats-accounting path
-    //! that previously sat at 0% coverage.
-
-    use std::time::Duration;
-
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-
-    use super::*;
-
-    /// Spawn a loopback TCP echo server, returning its `host:port`.
-    async fn spawn_echo_server() -> String {
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind echo server");
-        let addr = listener.local_addr().expect("addr").to_string();
-        tokio::spawn(async move {
-            while let Ok((mut sock, _)) = listener.accept().await {
-                tokio::spawn(async move {
-                    let mut buf = [0u8; 4096];
-                    loop {
-                        match sock.read(&mut buf).await {
-                            Ok(0) | Err(_) => break,
-                            Ok(n) => {
-                                if sock.write_all(&buf[..n]).await.is_err() {
-                                    break;
-                                }
-                            }
-                        }
-                    }
-                });
-            }
-        });
-        addr
-    }
-
-    fn free_local_addr() -> String {
-        let listener = std::net::TcpListener::bind("127.0.0.1:0").expect("bind");
-        let addr = listener.local_addr().expect("addr");
-        // Drop the listener so the port is free (nothing listening -> refused).
-        addr.to_string()
-    }
-
-    #[tokio::test]
-    async fn relays_channel_to_local_target_and_records_stats() {
-        let addr = spawn_echo_server().await;
-        let stats = Arc::new(ForwarderStats::new());
-
-        // `channel_stream` stands in for the SSH channel; `peer` is the remote
-        // side that would write into it.
-        let (channel_stream, mut peer) = tokio::io::duplex(64 * 1024);
-        let stats_clone = Arc::clone(&stats);
-        let addr_clone = addr.clone();
-        let relay = tokio::spawn(async move {
-            relay_to_local(channel_stream, &addr_clone, &stats_clone).await;
-        });
-
-        peer.write_all(b"hello").await.expect("write to channel");
-        let mut echoed = [0u8; 5];
-        peer.read_exact(&mut echoed).await.expect("read echo");
-        assert_eq!(&echoed, b"hello", "channel <-> local relay round-trips");
-
-        peer.shutdown().await.expect("half-close channel");
-        drop(peer);
-
-        tokio::time::timeout(Duration::from_secs(3), relay)
-            .await
-            .expect("relay finished")
-            .expect("relay task ok");
-
-        let snap = stats.to_tunnel_stats();
-        assert_eq!(snap.bytes_sent, 5);
-        assert_eq!(snap.bytes_received, 5);
-    }
-
-    #[tokio::test]
-    async fn relay_to_unreachable_local_target_is_a_no_op() {
-        let addr = free_local_addr(); // nothing is listening here
-        let stats = Arc::new(ForwarderStats::new());
-        let (channel_stream, _peer) = tokio::io::duplex(1024);
-
-        // Should return promptly without panicking and without recording bytes.
-        tokio::time::timeout(
-            Duration::from_secs(3),
-            relay_to_local(channel_stream, &addr, &stats),
-        )
-        .await
-        .expect("relay returns on connect failure");
-
-        let snap = stats.to_tunnel_stats();
-        assert_eq!(snap.bytes_sent, 0);
-        assert_eq!(snap.bytes_received, 0);
-    }
-}
+pub use termihub_core::tunnel::remote_forward::RemoteForwarder;
